1. 临时提交.

+++ xplat-meta-oqsengine-devops/src/main/java/com/xforceplus/ultraman/oqsengine/devops/rebuild/handler/AnyDevOpsTaskHandler.java
@@ -37,7 +37,7 @@ public class AnyDevOpsTaskHandler implements TaskHandler {
     }
 
     @Override
-    public int getProgressPercentage() {
+    public long getProgressPercentage() {
         return 0;
     }
