+++ xplat-meta-oqsengine-devops/src/main/java/com/xforceplus/ultraman/oqsengine/devops/rebuild/handler/DefaultDevOpsTaskHandler.java
@@ -72,7 +72,7 @@ public class DefaultDevOpsTaskHandler implements TaskHandler {
     }
 
     @Override
-    public int getProgressPercentage() {
+    public long getProgressPercentage() {
         if (!isDone()) {
             return devOpsTaskInfo.getProgressPercentage();
         }
